Added action_topic
Ref: https://www.home-assistant.io/integrations/climate.mqtt/#configuration

action_template template (Optional)
A template to render the value received on the action_topic with.

action_topic string (Optional)
The MQTT topic to subscribe for changes of the current action. If this is set, the climate graph uses the value received as data source. A “None” payload resets the current action state. An empty payload is ignored. Valid action values: off, heating, cooling, drying, idle, fan.